Move both chart components into a separate chart folder
