+++ src/components/charts/barChart/barChart-view.js
