+++ src/components/charts/barChart/index.js
