+++ src/components/charts/radarChart/radarChart-view.js
